pyproject.toml
[project]
name = "RAXO"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.2.0",
    "google-images-search>=1.4.0",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "nltk>=3.9.2",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pycocotools>=2.0.10",
    "python-dotenv>=1.2.1",
    "sam-2",
    "supervision>=0.27.0",
    "tabulate>=0.9.0",
    "tidecv>=1.0.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "wandb>=0.23.0",
    "weave>=0.52.17",
]

[tool.uv.sources]
sam-2 = { path = "../sam2" }
